--- src/idl2json/src/main.rs
@@ -1,17 +1,0 @@
-use candid::IDLArgs;
-use idl2json::idl_to_serde;
-use std::io::{self, Read};
-
-/// Reads IDL from stdin, writes JSON to stdout.
-fn main() -> io::Result<()> {
-    let mut buffer = String::new();
-    io::stdin().read_to_string(&mut buffer)?;
-
-    let args: IDLArgs = buffer.parse().expect("Malformed input");
-    println!(
-        "{}",
-        serde_json::to_string(&idl_to_serde(&args.args[0])).expect("Cannot get it out")
-    );
-
-    Ok(())
-}
