+++ src/idl2json_cli/Cargo.toml
@@ -5,4 +5,11 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[[bin]]
+name = "idl2json"
+path = "src/main.rs"
+
 [dependencies]
+candid = "0.6.21"
+serde_json = "^1.0"
+idl2json = { path = "../idl2json" }
